# src/openapi.yaml
openapi: 3.0.3
info:
  title: Gym Platform API
  version: 0.2.0
  description: API per palestre, piani, abbonamenti, prenotazioni e partner portal.
servers:
- url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Plan:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        plan_type:
          type: string
        price_cents:
          type: integer
        duration_days:
          type: integer
        entries_total:
          type: integer
paths:
  /health:
    get:
      summary: Health
      responses:
        '200':
          description: OK
  /gyms:
    get:
      summary: Cerca palestre
      parameters:
      - name: lat
        in: query
        schema:
          type: number
      - name: lon
        in: query
        schema:
          type: number
      - name: radius
        in: query
        schema:
          type: integer
          default: 2000
      - name: zone
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Lista
  /gyms/near/{gymId}:
    get:
      summary: Palestre vicine a gymId
      parameters:
      - name: gymId
        in: path
        required: true
        schema:
          type: integer
      - name: radius
        in: query
        schema:
          type: integer
          default: 2000
      responses:
        '200':
          description: Ok
  /gyms/{id}:
    get:
      summary: Dettaglio palestra
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Ok
  /plans:
    get:
      summary: Piani per palestra
      parameters:
      - name: gym_id
        in: query
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Ok
  /subscriptions:
    post:
      summary: Crea abbonamento
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Creato
  /bookings:
    post:
      summary: Crea prenotazione+QR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Creato
  /checkin/verify:
    post:
      summary: Verifica QR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Ok
  /payments/webhook:
    post:
      summary: Webhook pagamenti
      responses:
        '200':
          description: Ok
  /auth/partner/login:
    post:
      summary: Login partner
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Ok
  /partner/slots:
    get:
      summary: Slot per data (partner)
      security:
      - bearerAuth: []
      parameters:
      - name: gym_id
        in: query
        required: true
        schema:
          type: integer
      - name: date
        in: query
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Ok
  /partner/slots/{id}:
    patch:
      summary: Aggiorna slot (partner)
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Ok
  /partner/plans:
    post:
      summary: Crea piano (partner)
      security:
      - bearerAuth: []
      responses:
        '201':
          description: Creato
  /partner/plans/{id}:
    patch:
      summary: Aggiorna piano (partner)
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Ok
    delete:
      summary: Elimina piano (partner)
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Ok
  /partner/checkins:
    get:
      summary: Check-in giorno (partner)
      security:
      - bearerAuth: []
      parameters:
      - name: gym_id
        in: query
        required: true
        schema:
          type: integer
      - name: date
        in: query
        required: true
        schema:
          type: string
      - name: q
        in: query
        schema:
          type: string
      responses:
        '200':
          description: Ok